backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
SQLAlchemy
psycopg2-binary
WooCommerce